% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine CSV files}
\usage{
combine(data_folder = "data")
}
\arguments{
\item{data_folder}{Folder that CSV files are contained within}
}
\value{
Returns a data table of unique rows (removes duplicates)
}
\description{
Combine CSV files
}
\examples{
ATS_collar_data <- combine("data/ATS")
}
