% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ats.R
\name{convert_ATS}
\alias{convert_ATS}
\title{Convert ATS collar data into common format}
\usage{
convert_ATS(data = ATS_combined)
}
\arguments{
\item{data}{data frame containing raw ATS data. See combine function for merging csv files.}
}
\value{
Returns collar data converted into common format
}
\description{
Convert ATS collar data into common format
}
\examples{
  ATS_out <- convert_ATS(ATS_combined)
}
