% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_format.R
\name{convert_format}
\alias{convert_format}
\title{Convert collar data from multiple manufacturers into a common format}
\usage{
convert_format(data)
}
\arguments{
\item{data}{List of data frames to convert to a common format.  Automatically recognizes ATS and Vectronics based on dataframe names.}
}
\value{
Returns a single dataframe of combined and converted data
}
\description{
Convert collar data from multiple manufacturers into a common format
}
\examples{
converted_collardata <- convert_format(collardata)
}
