% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vec.R
\name{convert_vec}
\alias{convert_vec}
\title{Convert raw vectronics GPS collar data into common format}
\usage{
convert_vec(data = vec_combined)
}
\arguments{
\item{data}{Dataframe containing raw Vectronics data to be converted}
}
\value{
Return data frame of converted data in common format
}
\description{
Convert raw vectronics GPS collar data into common format
}
\examples{
  vec_out <- convert_vec(vec_combined)
}
