% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{latest}
\alias{latest}
\title{Return latest position for each animal ID from matric of GPS collar data}
\usage{
latest(data = data, range = 180)
}
\arguments{
\item{data}{Data frame with ID and GPS collar data}

\item{range}{Restrict data to within x number of days from current system time}
}
\description{
Return latest position for each animal ID from matric of GPS collar data
}
\examples{
latest_position <- latest(all_collars, range = 180)
}
