% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_database.R
\name{read_database}
\alias{read_database}
\title{Read latest local database}
\usage{
read_database(data_folder = NA, colClasses = NULL)
}
\arguments{
\item{data_folder}{Folder containing database to be read in.}
}
\value{
Copy of target database
}
\description{
Read latest local database
}
\examples{
read_database("data/ats")
}
