% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read multiple collar data files from computer hard drive}
\usage{
read_files(data_folder = "data", file_type = c("csv", "txt"), sep = ",")
}
\arguments{
\item{data_folder}{Folder location(s) of collar data files relative to parent directory}

\item{file_type}{Default is csv and txt}

\item{sep}{Delineator of data within files - default is ","}
}
\value{
Returns a list of dataframes, one for each folder
}
\description{
Read multiple collar data files from computer hard drive
}
\examples{
combined <- read.files(data_folder = c("data/Vectronics", "data/ATS", "data/database"))
}
