% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_database.R
\name{update_database}
\alias{update_database}
\title{Update database file and produce GPX of latest positions}
\usage{
update_database(range = 180)
}
\arguments{
\item{range}{Cut off number of days old for GPS positions when creating GPX file}
}
\description{
Update database file and produce GPX of latest positions
}
\examples{
update_database()
}
