% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gpx.R
\name{create_gpx}
\alias{create_gpx}
\title{Create GPX file from dataframe of GPS collar data}
\usage{
create_gpx(data = "latest_position", savelocation = "data/GPX")
}
\arguments{
\item{data}{Dataframe containing GPS point data}

\item{savelocation}{Location to save output GPX file within parent directory}
}
\value{
Create GPX file with latest points
}
\description{
Create GPX file from dataframe of GPS collar data
}
\examples{
create_gpx(latest_position)
}
